коррекция типов для выдачи голды командой

+++ src/server/scripts/Commands/cs_modify.cpp
@@ -1036,13 +1036,13 @@ public:
         if (handler->HasLowerSecurity(target, 0))
             return false;
 
-        int32 addmoney = atoi((char*)args);
+        int64 addmoney = atoi((char*)args);
 
         uint64 moneyuser = target->GetMoney();
 
         if (addmoney < 0)
         {
-            int64 newmoney = int32(moneyuser) + addmoney;
+            int64 newmoney = int64(moneyuser) + addmoney;
 
             sLog->outDebug(LOG_FILTER_GENERAL, handler->GetTrinityString(LANG_CURRENT_MONEY), moneyuser, addmoney, newmoney);
             if (newmoney <= 0)
